Create domain objects

+++ src/main/java/com/market/demo/persistence/entity/ComprasProductoPK.java
@@ -17,23 +17,23 @@ public class ComprasProductoPK implements Serializable {
 
 	private static final long serialVersionUID = -282515291384294805L;
 	@Column(name = "id_compra")
-	private Long idCompra;
+	private Integer idCompra;
 	@Column(name = "id_producto")
-	private Long idProducto;
+	private Integer idProducto;
 
-	public Long getIdCompra() {
+	public Integer getIdCompra() {
 		return idCompra;
 	}
 
-	public void setIdCompra(Long idCompra) {
+	public void setIdCompra(Integer idCompra) {
 		this.idCompra = idCompra;
 	}
 
-	public Long getIdProducto() {
+	public Integer getIdProducto() {
 		return idProducto;
 	}
 
-	public void setIdProducto(Long idProducto) {
+	public void setIdProducto(Integer idProducto) {
 		this.idProducto = idProducto;
 	}
